Multiple modules: extract deviceMemory / hardwareConcurrency to library, add codeQL warnings (#12070)
* Custom codeQL rules / hardwareConcurrency and deviceMemory

* move deviceMemory / hardwareConcurrency to a library

* reuse library code for deviceMemory & co
